requirements/base.txt
Django==1.6.2
South==0.8.4
distribute==0.7.3
django-allauth==0.16.1
django-extensions==1.3.3
django-floppyforms==1.1.1
django-reversion==1.8.0
django-compressor==1.3
django-polymorphic==0.5.3
django-suit==0.2.7
django-positions==0.5.1
oauthlib==0.6.1
psycopg2==2.5.2
python-openid==2.2.5
requests==2.2.1
requests-oauthlib==0.4.0
six==1.6.1
sorl-thumbnail==11.12.1b
wsgiref==0.1.2
Unicode==0.04.14
